fix(nuke): split shell-free reload listener so main.ts skips just-bash
Codex review noted that importing `installNukeReloadListener` from
`nuke-command.ts` made `main.ts` statically pull in `defineCommand`
from `just-bash` — ~800 kB of shell runtime — for every non-fixture
page boot. The page-side listener only needs the `BroadcastChannel`
+ key-clear logic, not the shell command itself.

Split the page-side wiring (`NUKE_CONTROL_CHANNEL`, `NukeReloadMsg`,
`NUKE_LOCAL_STORAGE_KEYS`, `installNukeReloadListener`) into a new,
zero-import `shell/supplemental-commands/nuke-channel.ts`.
`nuke-command.ts` re-exports those symbols so existing consumers and
the test suite keep their import paths, and the boot helper now
imports straight from `nuke-channel.ts` — main.ts no longer has any
path that reaches `defineCommand`.

Verified with the full @slicc/webapp suite (6428 passed, 10
skipped), `npm run lint:ci`, and `npm run typecheck`.
